vulnerability: update URL param
there was a filtering in update on Vulnerability side - Dashboard needs
to be switched as well. This is 1st commit which is backwards compatible
and also ensures nothing will break once Vulnerability updates
production.
This commit will be followed by another one once Vulnerability
production is updated which will remove the part used for compatability.