chore(RHIF-221): Update /beta to /preview (#359)
To test, you can run via `/beta` or `/preview`. We previously had issues with the local app being served. To ensure that it is properly being served, run the app via beta and make a change in the code or add a console log and ensure that you're getting local changes.

Co-authored-by: Michael Johnson <micjohns@newbeginnings.usersys.redhat.com>