fix(RHINENG-17866): Fix buttons on loading
This PR fixes an issue where the "View CVEs" and "View known exploits"
buttons don't have the filters they need because the data hasn't fully
loaded. The `affectingFilterValue` is set to 'true' during the loading
state, and if you click the buttons before the card is loaded, it will
cause an error on the vulnerabilities tables.

This PR converts the buttons to a Skeleton on loading state so they
cannot be clicked.

To test, go to `/insights/dashboard` and see that the "View CVEs" and
"View known exploits" buttons are Skeletons on loading. Also, test that
you cannot click them.