fix(RHINENG-21066): add filter to patch sys link
This adds a filter in the affected systems link in the Patch dashboard
to prevent a confusing system count mismatch. The link was just
redirecting to the systems page, which by default includes fresh and
stale systems, hence the number of systems shown could be different from
the one in the dashboard as that doesn't take stale ones into account.